requirements.txt
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
nltk==3.9.2
fastapi==0.121.0
uvicorn==0.38.0
python-dotenv==1.2.1
google-generativeai==0.8.5
langchain-google-genai==3.0.1
langchain-core==1.0.3
requests==2.32.5
tqdm==4.67.1